requirements.txt
drake==1.3
lxml
matplotlib
networkx
notebook
numpy
pandas
pydot
scipy

# Adds this project to path
-e .
